@@ -1,4 +1,4 @@
-﻿@inherits BlazorLayoutComponent
+@inherits LayoutComponentBase
 
 <NavMenu />
 
