@@ -1,4 +1,4 @@
-﻿@page "/blocks"
+@page "/blocks"
 @page "/blocks/{blockName}"
 @inject BlockRepository BlockRepository
 
@@ -31,25 +31,25 @@
     }
 </section>
 
-@functions {
+@code {
     private UnicodeBrowser.Client.Models.BlockInformation[] Blocks { get; set; }
 
     private string _blockName;
 
     [Parameter]
-    private string BlockName
+    public string BlockName
     {
         get => _blockName;
         set
         {
-            _blockName = value != null ? Uri.UnescapeDataString(value) : null; // Workaround for Blazor not unescaping route segments
+            _blockName = value;
             MapSelectedBlock();
         }
     }
 
     private UnicodeBrowser.Client.Models.BlockInformation SelectedBlock { get; set; }
 
-    protected override async Task OnInitAsync()
+    protected override async Task OnInitializedAsync()
     {
         Blocks = await BlockRepository.GetBlocksAsync();
         MapSelectedBlock();
