@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Blazor;
+using Microsoft.AspNetCore.Blazor;
+using Microsoft.AspNetCore.Components;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
